Додав аксесконтроль для функцій в фото

+++ server/src/routes/photos.py
@@ -1,10 +1,11 @@
-from src.database.models import User, Picture
+from src.database.models import User, Picture, Role
 from src.repository.photos import PictureRepository
 from fastapi import UploadFile, File, Form
 import logging
 from fastapi import APIRouter, HTTPException, Query, Depends
 from typing import Optional, List
 from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.future import select
 from src.database.db import get_db
 from src.services.auth import auth_service
 from src.schemas.photos import PictureUpload, PictureResponse
@@ -44,7 +45,8 @@ async def search_pictures(
     :rtype: List[PictureResponse]
     """
 
-    pictures = await PictureRepository.search_pictures(db=db, search_term=search_term, tag=tag, user_id=user_id, page=page, page_size=page_size)
+    pictures = await PictureRepository.search_pictures(db=db, search_term=search_term, tag=tag, user_id=user_id,
+                                                       page=page, page_size=page_size)
 
     return pictures
 
@@ -138,7 +140,6 @@ async def update_picture(
     return picture
 
 
-
 @router.delete("/{picture_id}", response_model=PictureResponse)
 async def delete_picture(picture_id: int, db: AsyncSession = Depends(get_db),
                          current_user: User = Depends(auth_service.get_current_user)):
@@ -159,6 +160,12 @@ async def delete_picture(picture_id: int, db: AsyncSession = Depends(get_db),
     Raises:
     HTTPException: If the picture is not found.
     """
+
+    picture = await db.execute(select(Picture).filter(Picture.id == picture_id))
+    picture = picture.scalar()
+    if picture.user_id != current_user.id or current_user.role.name != 'admin':
+        raise HTTPException(status_code=403, detail="Not enough permissions")
+
     picture = await PictureRepository.delete_picture(picture_id, db)
     if not picture:
         raise HTTPException(status_code=404, detail="Picture not found")
@@ -235,7 +242,6 @@ async def overlay_image(
     return picture
 
 
-  
 @router.get("/{picture_id}/tags", response_model=List[str])
 async def get_tags(
 
@@ -243,7 +249,6 @@ async def get_tags(
         db: AsyncSession = Depends(get_db),
         current_user: User = Depends(auth_service.get_current_user)
 ):
-
     """
     Route handler for retrieving the tags of a specific picture.
 
@@ -264,11 +269,15 @@ async def get_tags(
     return tags
 
 
-  
 @router.post("/{picture_id}/qrcode", response_model=PictureResponse)
 async def create_qrcode(picture_id: int, db: AsyncSession = Depends(get_db),
-                     current_user: User = Depends(auth_service.get_current_user)):
-  
+                        current_user: User = Depends(auth_service.get_current_user)):
+
+    picture = await db.execute(select(Picture).where(Picture.id == picture_id).first())
+    picture = picture.scalar()
+    if picture.user_id != current_user.id or current_user.role.name != 'admin':
+        raise HTTPException(status_code=403, detail="Not enough permissions")
+
     picture = await PictureRepository.create_qrcode(picture_id, db)
     if not picture:
         raise HTTPException(status_code=404, detail="Picture not found")
@@ -284,4 +293,3 @@ async def get_qrcode(picture_id: int, db: AsyncSession = Depends(get_db),
         raise HTTPException(status_code=404, detail="Picture not found")
 
     return picture.qr_code_url
-
